html/parsing/manakai/foreign-attrs-1.dat

#data
<!DOCTYPE HTML><svg xml:Lang="abc"/>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       xml lang="abc"

#data
<!DOCTYPE HTML><svg xml:space="abc"/>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       xml space="abc"

#data
<!DOCTYPE HTML><svg xml:base="abc"/>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       xml:base="abc"

#data
<!DOCTYPE HTML><svg xml:ID="abc"/>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       xml:id="abc"

#data
<!DOCTYPE HTML><svg xml:fuga="abc"/>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       xml:fuga="abc"

#data
<!DOCTYPE HTML><math xml:Lang="abc"/>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       xml lang="abc"

#data
<!DOCTYPE HTML><math xml:space="abc"/>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       xml space="abc"

#data
<!DOCTYPE HTML><math xml:base="abc"/>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       xml:base="abc"

#data
<!DOCTYPE HTML><math xml:ID="abc"/>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       xml:id="abc"

#data
<!DOCTYPE HTML><math xml:fuga="abc"/>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       xml:fuga="abc"

#data
<!DOCTYPE HTML><svg><g xml:Lang="abc"/></svg>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         xml lang="abc"

#data
<!DOCTYPE HTML><svg><g xml:space="abc"/></svg>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         xml space="abc"

#data
<!DOCTYPE HTML><svg><g xml:base="abc"/></svg>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         xml:base="abc"

#data
<!DOCTYPE HTML><svg><g xml:ID="abc"/></svg>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         xml:id="abc"

#data
<!DOCTYPE HTML><svg><g xml:fuga="abc"/></svg>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         xml:fuga="abc"

#data
<!DOCTYPE html><p xml:lang="a">
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       xml:lang="a"

#data
<!DOCTYPE html><p xml:space="a">
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       xml:space="a"

#data
<!DOCTYPE html><p xml:base="a">
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       xml:base="a"

#data
<!DOCTYPE html><p xml:id="a">
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       xml:id="a"
